requirement.txt
numpy                   1.24.3
opencv-python-headless  4.5.3.56
Pillow                  9.2.0
scikit-learn            1.3.2
scipy                   1.10.1
segment-anything        1.0    
tensorboard             2.14.0
torch                   1.7.1
torchaudio              0.7.2
torchvision             0.16.1